+++ src/DevChatter.Bot.Core/Commands/IBotCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using DevChatter.Bot.Core.Data.Model;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 
 namespace DevChatter.Bot.Core.Commands
