+++ src/DevChatter.Bot.Core/Events/Args/CommandReceivedEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using DevChatter.Bot.Core.Data.Model;
 
-namespace DevChatter.Bot.Core.Events
+namespace DevChatter.Bot.Core.Events.Args
 {
     public class CommandReceivedEventArgs
     {
