+++ src/DevChatter.Bot.Core/Events/Args/NewSubscriberEventArgs.cs
@@ -1,4 +1,4 @@
-﻿namespace DevChatter.Bot.Core.Events
+﻿namespace DevChatter.Bot.Core.Events.Args
 {
     public class NewSubscriberEventArgs
     {
