+++ src/DevChatter.Bot.Core/Events/Args/UserNoticedEventArgs.cs
@@ -1,6 +1,6 @@
 ﻿using DevChatter.Bot.Core.Data.Model;
 
-namespace DevChatter.Bot.Core.Events
+namespace DevChatter.Bot.Core.Events.Args
 {
     public class UserStatusEventArgs
     {
