--- src/DevChatter.Bot.Core/Events/CommandAliasChangedEventArgs.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DevChatter.Bot.Core.Events
-{
-    class CommandAliasChangedEventArgs : EventArgs
-    {
-    }
-}
