+++ src/DevChatter.Bot.Core/Events/SubscriberHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 
 namespace DevChatter.Bot.Core.Events
