+++ src/DevChatter.Bot.Core/Games/RockPaperScissors/RockPaperScissorsCommand.cs
@@ -3,6 +3,7 @@ using DevChatter.Bot.Core.Commands;
 using DevChatter.Bot.Core.Data;
 using DevChatter.Bot.Core.Data.Model;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 
 namespace DevChatter.Bot.Core.Games.RockPaperScissors
