+++ src/DevChatter.Bot.Core/Messaging/Tokens/SimpleToken.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 
 namespace DevChatter.Bot.Core.Messaging.Tokens
 {
