+++ src/DevChatter.Bot.Core/Systems/Chat/IChatClient.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using DevChatter.Bot.Core.Data.Model;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 
 namespace DevChatter.Bot.Core.Systems.Chat
 {
