+++ src/DevChatter.Bot.Core/Systems/Streaming/FollowableSystem.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 
 namespace DevChatter.Bot.Core.Systems.Streaming
