+++ src/DevChatter.Bot.Core/Systems/Streaming/IFollowerService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 
 namespace DevChatter.Bot.Core.Systems.Streaming
 {
