+++ src/DevChatter.Bot.Infra.Twitch/TwitchChatClient.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using DevChatter.Bot.Core.Data.Model;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 using DevChatter.Bot.Infra.Twitch.Extensions;
 using TwitchLib;
