+++ src/DevChatter.Bot/Startup/SetUpBot.cs
@@ -11,6 +11,7 @@ using DevChatter.Bot.Infra.Twitch;
 using DevChatter.Bot.Infra.Twitch.Events;
 using System.Collections.Generic;
 using System.Linq;
+using DevChatter.Bot.Core.Commands.Trackers;
 using DevChatter.Bot.Core.Util;
 using TwitchLib;
