+++ src/UnitTests/Core/Commands/QuoteCommandTests/ProcessShould.cs
@@ -5,6 +5,7 @@ using DevChatter.Bot.Core.Data;
 using DevChatter.Bot.Core.Data.Model;
 using DevChatter.Bot.Core.Data.Specifications;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 using Moq;
 using Xunit;
