+++ src/UnitTests/Core/Commands/SimpleCommandTests/ProcessShould.cs
@@ -1,5 +1,6 @@
 ﻿using DevChatter.Bot.Core.Commands;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 using Moq;
 using UnitTests.Fakes;
