+++ src/UnitTests/Core/Events/CommandHandlerTests/CommandReceivedHandlerShould.cs
@@ -1,9 +1,11 @@
 ﻿using System.Collections.Generic;
 using DevChatter.Bot.Core.Commands;
+using DevChatter.Bot.Core.Commands.Trackers;
 using DevChatter.Bot.Core.Data;
 using DevChatter.Bot.Core.Data.Model;
 using DevChatter.Bot.Core.Data.Specifications;
 using DevChatter.Bot.Core.Events;
+using DevChatter.Bot.Core.Events.Args;
 using DevChatter.Bot.Core.Systems.Chat;
 using Moq;
 using UnitTests.Fakes;
